early stage working module
this is a early stage module that is succesfully working and reading the data. For now we are still working on how to use the data in order to detect schizophrenia in the paitents. using the convert code, we convert the CSV format of the data into a FIF file formart which can be read by the egg_pipeline code. For the sample size is small, and we are working on increasing the sample size as well as trying to figure out how to use this data and detect the condition.